Remove the _internals.comments module. Fix pyrefly check when running.
